MNT: move to obj_config module, slight logic touchups
